+++ load.py
@@ -17,6 +17,8 @@ from config import appname, config
 from theme import theme
 from ttkHyperlinkLabel import HyperlinkLabel
 
+from ScrollableNotebook import *
+
 this = sys.modules[__name__]  # For holding module globals
 this.VersionNo = "1.9.0"
 this.GitVersion = "0.0.0"
@@ -588,7 +590,8 @@ def display_data(title, data, tick_mode):
 
     ContainerFrame = ttk.Frame(Form)
     ContainerFrame.pack(fill=tk.BOTH, expand=1)
-    TabParent = ttk.Notebook(ContainerFrame)
+    TabParent=ScrollableNotebook(ContainerFrame, wheelscroll=False, tabmenu=True)
+    #TabParent = ttk.Notebook(ContainerFrame)
     TabParent.pack(fill=tk.BOTH, expand=1, side=tk.TOP, padx=5, pady=5)
 
     DiscordFrame = ttk.Frame(ContainerFrame)
@@ -610,15 +613,23 @@ def display_data(title, data, tick_mode):
     ttk.Checkbutton(DiscordOptionsFrame, text="Abbreviate Faction Names", variable=this.AbbreviateFactionNames, onvalue=CheckStates.STATE_ON, offvalue=CheckStates.STATE_OFF, command=partial(option_change, Discord, data)).grid(sticky=tk.W)
     ttk.Checkbutton(DiscordOptionsFrame, text="Include Secondary INF", variable=this.IncludeSecondaryInf, onvalue=CheckStates.STATE_ON, offvalue=CheckStates.STATE_OFF, command=partial(option_change, Discord, data)).grid(sticky=tk.W)
 
+    # Calculate whether each system has had any activity and store in data structure
     for i in data:
         z = len(data[i][0]['Factions'])
-        zero_system_activity = True
+        data[i][0]['zero_system_activity'] = True
         for x in range(0, z):
             update_faction_data(data[i][0]['Factions'][x])
             if not is_faction_data_zero(data[i][0]['Factions'][x]):
-                zero_system_activity = False
+                data[i][0]['zero_system_activity'] = False
+                break
 
-        if this.ShowZeroActivitySystems.get() == CheckStates.STATE_OFF and zero_system_activity:
+    # Create a list of integer indexes into the data object, sorted by System name, prioritising systems with activity first
+    sorted_data_indexes = sorted(data, key=lambda x: ("1_" if data[x][0]['zero_system_activity'] else "0_") + data[x][0]['System'])
+
+    for i in sorted_data_indexes:
+        z = len(data[i][0]['Factions'])
+
+        if this.ShowZeroActivitySystems.get() == CheckStates.STATE_OFF and data[i][0]['zero_system_activity']:
             continue
 
         tab = ttk.Frame(TabParent)
@@ -727,6 +738,8 @@ def display_data(title, data, tick_mode):
                 CZGroundHVar.trace('w', partial(cz_change, CZGroundHVar, Discord, CZs.GROUND_HIGH, data, i, x))
 
         update_enable_all_factions_checkbutton(EnableAllCheckbutton, FactionEnableCheckbuttons)
+
+        tab.pack_forget()
 
     Discord.insert(tk.INSERT, generate_discord_text(data))
     # Select all text and focus the field